A righeous set of renamings
