+++ metafiddler/input.py
@@ -5,35 +5,31 @@
 
 # I mean, this could be REST or captive KBD or who knows.
 # Input however should be nonblocking.
-import metafiddler.controller_implementation.windows_keyboard
-import metafiddler.controller_implementation.windows_usb_joystick
-import Event
+import metafiddler.controller.windows_keyboard
+import metafiddler.controller.windows_usb_joystick
+from metafiddler.events.input import Event
 
 # TODO: feels janky, load from files mebbe?
 controllers = [
-   metafiddler.controller_implementation.windows_keyboard,
-   metafiddler.controller_implementation.windows_usb_joystick
+   metafiddler.controller.windows_keyboard,
+   metafiddler.controller.windows_usb_joystick
 ]
 
 class Input:
     # Maybe debouncing here is the more humane thing to do?  Of course it is...
     last_events = [Event.NONE, Event.NONE]
 
-    def __init(self)__:
-        global last_event
-        global controllers
+    def __init__(self):
         for controller in controllers:
             controller.init()
 
     def poll(self):
-        global last_event
-        global controllers
         #for controller in controllers:
         for x in range(len(controllers)):
             controller = controllers[x]
             event = controller.poll()
-            if last_events[x] != event:
-                last_events[x] = event
+            if self.last_events[x] != event:
+                self.last_events[x] = event
                 return event
 
         return Event.NONE
